pyproject.toml
[project]
name = "funding-service-pre-award"
version = "0.1.0"
description = "The Funding Service pre-award frontends and stores."
readme = "README.md"
license = "MIT License"
requires-python = ">=3.10, <3.11"

dependencies = [
    "airium==0.2.6",
    "alembic-utils==0.8.5",
    "babel==2.16.0",
    "beautifulsoup4==4.12.3",
    "boto3==1.35.96",
    "colored==2.2.4",
    "cssmin==0.2.0",
    "dataclass-dict-convert==1.7.4",
    "email-validator==2.2.0",
    "flask-assets==2.1.0",
    "flask-babel==4.0.0",
    "flask-compress==1.15",
    "flask-json==0.4.0",
    "flask-marshmallow==1.2.1",
    "flask-migrate==4.0.7",
    "flask-redis==0.4.0",
    "flask-restx==1.3.0",
    "flask-session==0.8.0",
    "flask-sqlalchemy==3.1.1",
    "flask-talisman==1.1.0",
    "flask-wtf==1.2.2",
    "flask==3.0.3",
    "funding-service-design-utils[toggles]==6.0.1",
    "govuk-frontend-jinja==2.7.0",
    "greenlet==3.1.1",
    "gunicorn[gevent]==23.0.0",
    "invoke==2.2.0",
    "jsmin==3.0.1",
    "jsonpath-rw-ext==1.2.2",
    "marshmallow-enum==1.5.1",
    "marshmallow-sqlalchemy==1.0.0",
    "msal==1.28.0",
    "num2words==0.5.14",
    "openapi-spec-validator==0.7.1",
    "openpyxl==3.1.5",
    "pandas==2.2.2",
    "prance==23.6.21.0",
    "psycopg2-binary==2.9.9",
    "pyjwt==2.8.0",
    "python-dateutil==2.9.0.post0",
    "python-slugify==8.0.4",
    "requests==2.32.3",
    "shortuuid==1.0.13",
    "slugify==0.0.1",
    "sqlalchemy-json==0.7.0",
    "sqlalchemy-utils==0.41.2",
    "sqlalchemy[mypy]==2.0.37",
    "xhtml2pdf==0.2.16",
    "notifications-python-client==10.0.1",
]

[tool.djlint]
# run with : `djlint path/to/file.html --reformat --format-css --format-js`
#   this is deliberately commented out.  we don't want to format these tags as
#   it will introduce new lines and tabs, making the translation matching brittle.
# custom_blocks="trans,endtrans"
max_line_length=1000        # high limit, we don't want line breaks for translations.
max_attribute_length=1000   # ^^^
exclude=".venv,venv"
profile="jinja2"

[tool.ruff]
line-length = 120

target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "W",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C90",  # mccabe cyclomatic complexity
    "G",  # flake8-logging-format
]
ignore = []
exclude = [
    "pre_award/db/migrations/versions/",
    "venv*",
    ".venv*",
    "__pycache__",
    "pre_award/fund_store/config/fund_loader_config/FAB/*"
]
mccabe.max-complexity = 12

[tool.uv]

[dependency-groups]
dev = [
    "asserts==0.13.1",
    "beautifulsoup4==4.12.3",
    "click==8.1.7",
    "colored==2.2.4",
    "debugpy==1.8.11",
    "deepdiff==8.1.1",
    "dparse==0.6.4",
    "invoke==2.2.0",
    "invoke==2.2.0",
    "json2html==1.3.0",
    "moto[s3,sqs]==5.0.26",
    "pre-commit==4.0.1",
    "pytest-env==1.1.5",
    "pytest-flask==1.3.0",
    "pytest-html==4.1.1",
    "pytest-mock==3.14.0",
    "pytest==8.3.4",
    "requests-mock==1.11.0",
    "ruff==0.9.1",
    "flask-debugtoolbar==0.16.0",
    "freezegun==1.5.1",
    "responses==0.25.3",
]
